feat: eslint fixes, added disable, title and deprecated disabled
