+++ src/atoms/button.scss
@@ -34,14 +34,19 @@ $buttonSize: 32px;
         }
     }
     &.button-color- {
+        &light, &outline, &transparent {
+            color: color(color);
+        }
         &alert { background-color: color(backgroundColorButtonAlert); }
         &alternate { background-color: color(backgroundColorButtonAlternate); }
-        &disable { background-color: color(backgroundColorButtonDisable); cursor: default; }
-        &light { background-color: color(backgroundColorLight); color: color(color); }
+        &disable {
+            background-color: color(backgroundColorButtonDisable);
+            cursor: default;
+        }
+        &light { background-color: color(backgroundColorLight); }
         &outline {
             background-color: transparent;
             box-shadow: inset 0 0 0 1px color(borderColorQuaternary);
-            color: color(color);
             > .ui.icon {
                 .icon-use-path { fill: color(color); }
             }
@@ -128,6 +133,9 @@ $buttonSize: 32px;
         }
     }
     &.button-inverse.button-color- {
+        &outline, &transparent {
+            color: color(colorInverse);
+        }
         &alert { background-color: color(backgroundColorButtonAlert); }
         &alternate { background-color: color(backgroundColorButtonAlternate); }
         &disable { background-color: color(backgroundColorButtonDisable); cursor: default;  }
@@ -135,7 +143,6 @@ $buttonSize: 32px;
         &outline {
             background-color: rgba(255, 255, 255, .30);
             box-shadow: inset 0 0 0 1px color(borderColorSecondary);
-            color: color(colorInverse);
             > .ui.icon {
                 .icon-use-path { fill: color(color); }
             }
@@ -143,7 +150,9 @@ $buttonSize: 32px;
         &primary { background-color: color(backgroundColorButton); }
         &success { background-color: color(backgroundColorButtonSuccess); }
         &secondary { background-color: color(backgroundColorButtonSecondary); }
-        &transparent { background-color: transparent; }
+        &transparent {
+            background-color: transparent;
+        }
         &warning { background-color: color(backgroundColorWarning); }
     }
     &.button-icon {
